Move SchemaId and DatabaseId
